AI Chat: allow copy pasting media files
