+++ apps/src/aiComponentLibrary/userMessageEditor/UserMessageEditor.tsx
@@ -25,6 +25,7 @@ export interface UserMessageEditorProps {
   editorContainerClassName?: string;
   customPlaceholder?: string;
   speechToTextEnabled?: boolean;
+  onPaste?: (event: React.ClipboardEvent<HTMLTextAreaElement>) => void;
   children?: React.ReactNode;
 }
 
@@ -44,6 +45,7 @@ const UserMessageEditor = React.forwardRef<
       customPlaceholder,
       showSubmitLabel = false,
       speechToTextEnabled = false,
+      onPaste,
       children,
     },
     externalInputRef
@@ -124,6 +126,7 @@ const UserMessageEditor = React.forwardRef<
           aria-label={commonI18n.aiUserMessagePlaceholder()}
           onFocus={() => setFocused(true)}
           onBlur={() => setFocused(false)}
+          onPaste={onPaste}
         />
         <div className={moduleStyles.chatActionsContainer}>
           {children}
