+++ apps/src/aiComponentLibrary/userMessageEditor/speechToTextButton/styles.module.scss
@@ -18,6 +18,10 @@
 .buttonContainer {
   position: relative;
   display: flex;
+
+  .microphoneButtonContainer {
+    display: flex;
+  }
 }
 
 @keyframes ping {
