+++ apps/src/aichat/redux/thunks/index.ts
@@ -11,3 +11,4 @@ export {submitTeacherFeedback} from './submitTeacherFeedback';
 export {updateAiCustomization} from './updateAiCustomization';
 export {fetchThreadMessages} from './fetchThreadMessages';
 export {initializeAiCustomizations} from './initializeAiCustomizations';
+export {uploadFiles} from './uploadFiles';
