+++ apps/src/aichat/views/UserChatMessageEditor.tsx
@@ -7,7 +7,12 @@ import experiments from '@cdo/apps/util/experiments';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
 import supportsClientApi from '../api/supportsClientApi';
-import {selectIsWaitingForChatResponse, submitChatContents} from '../redux';
+import {ACCEPTED_FILE_TYPES} from '../constants';
+import {
+  selectIsWaitingForChatResponse,
+  submitChatContents,
+  uploadFiles,
+} from '../redux';
 import {
   AiChatClientType,
   ChatButtonAndKey,
@@ -144,6 +149,22 @@ const UserChatMessageEditor: React.FunctionComponent<
     supportsClientApi(modelParameters.selectedModelId) ||
     experiments.isEnabledAllowingQueryString('enable-speech-to-text');
 
+  const onPaste = useCallback(
+    (e: React.ClipboardEvent<HTMLTextAreaElement>) => {
+      if (!multimodalAvailable || !buildAssetUrl) {
+        return;
+      }
+      const files = Array.from(e.clipboardData.items)
+        .filter(({type}) =>
+          ACCEPTED_FILE_TYPES.includes(`.${type.split('/')[1]}`)
+        )
+        .map(item => item.getAsFile())
+        .filter(item => item !== null);
+      dispatch(uploadFiles({files, buildAssetUrl}));
+    },
+    [multimodalAvailable, buildAssetUrl, dispatch]
+  );
+
   return (
     <>
       {chatButtons && chatButtons.length > 0 && !chatDisabled && (
@@ -160,6 +181,7 @@ const UserChatMessageEditor: React.FunctionComponent<
         disabled={disabled}
         editorContainerClassName={editorContainerClassName}
         speechToTextEnabled={speechToTextEnabled}
+        onPaste={onPaste}
         ref={inputRef}
       >
         {multimodalAvailable && buildAssetUrl && levelName && (
