+++ apps/src/aichat/views/assets/UploadButton.tsx
@@ -7,27 +7,15 @@ import {
 } from '@mui/material';
 import React, {ChangeEvent, useState} from 'react';
 
-import Lab2Registry from '@cdo/apps/lab2/Lab2Registry';
 import StarterAssetsDialog from '@cdo/apps/lab2/views/components/starterAssetsDialog';
 import {AssetData} from '@cdo/apps/lab2/views/components/starterAssetsDialog/types';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 import useHiddenFileInput from '@cdo/apps/util/hooks/useHiddenFileInput';
-import HttpClient, {NetworkError} from '@cdo/apps/util/HttpClient';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
-import {
-  ACCEPTED_FILE_TYPES,
-  MAX_FILE_SIZE_MB,
-  MAX_NUM_FILES,
-} from '../../constants';
+import {ACCEPTED_FILE_TYPES, MAX_NUM_FILES} from '../../constants';
 import aichatI18n from '../../locale';
-import {
-  addStagedFile,
-  clearStagedFilesAlert,
-  sendAnalytics,
-  stagedFilesLimitExceeded,
-  stagedFileUploadFinished,
-} from '../../redux';
+import {addStagedFile, sendAnalytics, uploadFiles} from '../../redux';
 import {AssetSource, ChatAsset} from '../../types';
 
 export interface UploadButtonProps {
@@ -57,92 +45,7 @@ const UploadButton: React.FC<UploadButtonProps> = ({
     if (!files) {
       return;
     }
-
-    // Clear the alert, if any.
-    dispatch(clearStagedFilesAlert());
-
-    const excessFileCount = files.length - numAllowedFiles;
-    if (excessFileCount > 0) {
-      dispatch(stagedFilesLimitExceeded());
-    }
-
-    const allowedFiles = Array.from(files)
-      .slice(0, numAllowedFiles)
-      .map<[string, ChatAsset, File]>(file => [
-        // Create a unique key for each upload in case the same file is uploaded more than once.
-        `${file.name}-${Date.now()}`,
-        {filename: file.name, source: AssetSource.PROJECT},
-        file,
-      ]);
-
-    for (const [key, asset] of allowedFiles) {
-      dispatch(addStagedFile({key, asset}));
-    }
-
-    let uploadSuccessCount = 0;
-    let sizeLimitExceededCount = 0;
-    let uploadFailureCount = 0;
-    let fileCountPdf = 0;
-    let fileCountImage = 0;
-    for (const [key, asset, file] of allowedFiles) {
-      if (file.size > MAX_FILE_SIZE_MB * 1024 * 1024) {
-        sizeLimitExceededCount += 1;
-        dispatch(
-          stagedFileUploadFinished({
-            key,
-            status: 'sizeLimitExceeded',
-          })
-        );
-        continue; // Skip uploading this file if it exceeds the size limit.
-      }
-
-      if (file.name.endsWith('.pdf')) {
-        fileCountPdf += 1;
-      } else {
-        fileCountImage += 1;
-      }
-
-      try {
-        await HttpClient.put(buildAssetUrl(asset), file);
-        uploadSuccessCount += 1;
-
-        dispatch(stagedFileUploadFinished({key, status: 'uploaded'}));
-      } catch (error) {
-        let status: 'sizeLimitExceeded' | 'uploadFailed' = 'uploadFailed';
-        if (error instanceof NetworkError && error.response.status === 413) {
-          sizeLimitExceededCount += 1;
-          status = 'sizeLimitExceeded';
-        } else {
-          uploadFailureCount += 1;
-          status = 'uploadFailed';
-          // Only log if not a size limit exceeded error.
-          Lab2Registry.getInstance()
-            .getMetricsReporter()
-            .logError('Error uploading asset', error as Error, {
-              filename: file.name,
-            });
-        }
-
-        dispatch(
-          stagedFileUploadFinished({
-            key,
-            status,
-          })
-        );
-      }
-    }
-
-    dispatch(
-      sendAnalytics(EVENTS.AICHAT_MULTIMODAL_UPLOAD_STAGED, {
-        source: AssetSource.PROJECT,
-        fileCountSuccess: uploadSuccessCount,
-        fileCountFailureSizeLimitExceeded: sizeLimitExceededCount,
-        fileCountFailureUnknownCause: uploadFailureCount,
-        fileCountFailureNumberExceeded: Math.max(excessFileCount, 0),
-        fileCountImage,
-        fileCountPdf,
-      })
-    );
+    dispatch(uploadFiles({files: Array.from(files), buildAssetUrl}));
   };
 
   const onSelectStarterAssets = (assets: AssetData[]) => {
